% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{MyFirstPackage}
\alias{MyFirstPackage}
\alias{MyFirstPackage-package}
\title{Plotting download numbers from R Studio CRAN Mirror}
\description{
MyFirstPackage is a test package which includes a function to plot the download numbers from the R Studio CRAN Mirror.
}
\references{
Your submitted paper which is in revision for the last couple of months.
}

