% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotDownloads}
\alias{plotDownloads}
\title{Plotting R Package Downloads by month or day.}
\usage{
plotDownloads(PackageName = "ggplot2", from = "2016-04-01",
  to = "2016-11-30", by = "month")
}
\arguments{
\item{PackageName}{Name of the package.}

\item{from}{Start date, in \code{yyyy-mm-dd} format.}

\item{to}{End date, in \code{yyyy-mm-dd} format.}

\item{by}{Plot downloads by \code{"day"} or \code{"month"}. It is advised to only use \code{"day"} for shorter periods of time to 
avoid too much clutter.}
}
\value{
A \code{\link{ggplot2}} object based on the parameter settings.
}
\description{
This functions creates a ggplot2 graph of the number of downloads of a chosen package from the R Studio CRAN mirror. 
The graph can be by month or by day.
}
\details{
When plotting by day, each month/year combination is assigned a different colour.
}
\examples{
\dontrun{
plotDownloads("ggplot2",from="2015-01-01",to="2016-11-30")
plotDOwnloads("ggplot2",from="2016-04-01",to="2016-11-30",by="day")
}
}
\author{
Ewoud De Troyer
}
\references{
Not really.
}

