% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TidyVerse_gplots}
\alias{TidyVerse_gplots}
\title{Tidyverse Plots}
\format{A list with 4 ggplot objects.}
\source{
R script in \code{doc/} folder
}
\description{
A \code{.RData} object which contains a couple of download plots of some packages of the Hadley Tidyverse.
}

